+++ src/components/General.styled.ts
@@ -17,3 +17,7 @@ export const Button = styled.button`
   margin-top: 50px;
   box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
 `;
+
+export const Span = styled.span`
+  color: #ff5300;
+`;
